make graph data one-class

+++ scripts/makeOneclass.sh
@@ -1,8 +1,10 @@
 #!/bin/sh
 
 for file in $1/*txt; do
-	echo $file
-	sed 's/^./0/g' "$file" > "$file.out"
-	rm "$file"
-	mv "$file.out" "$file"
+	if ! [[ "$file" == *"classes.txt" ]]; then
+		echo $file
+		sed 's/^./0/g' "$file" > "$file.out"
+		rm "$file"
+		mv "$file.out" "$file"
+	fi
 done;
